requirements_azureml.txt
PyYAML
azure-ai-ml~=1.24.0
azure-identity~=1.19.0
